# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local Scientific Event Manager System",
        "title": "Scientific Event Manager API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/event/regtype/create": {
            "post": {
                "description": "An event organizer can create a registration type for one of their unpublished events",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Create a Registration Type of an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the event",
                        "name": "eventID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the registration type",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Description of the registration type",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Price of the registration type",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of benefits",
                        "name": "benefits",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event Registration Type added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Can't parse the price",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "User is not the orgaziner of the event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Can't add a registration type to a published event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error found during registration type creation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/regtype/delete": {
            "post": {
                "description": "An event organizer can delete a registration type for one of their unpublished events",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Delete a Registration Type of an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the event",
                        "name": "eventID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the registration type",
                        "name": "regTypeID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event Registration Type deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "User is not the orgaziner of the event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Can't delete the registration type of a published event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error found during registration type deletion",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/regtype/edit": {
            "post": {
                "description": "An event organizer can edit a registration type for one of their unpublished events",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Edit a Registration Type of an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the event",
                        "name": "eventID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the registration type",
                        "name": "regTypeID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the registration type",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Description of the registration type",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Price of the registration type",
                        "name": "price",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of benefits",
                        "name": "benefits",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event Registration Type edited successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Can't parse the price",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "User is not the organizer of the event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Can't edit the registration type of a published event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error found during registration type creation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/event/view/:id/benefit_participants/:benefitID": {
            "get": {
                "description": "A event organizer can view a list of all the users eligible for a benefit given in one of their events.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "List the users eligible for a benefit in a event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the event",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the benefit",
                        "name": "benefitID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.ParticipantInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "User isn't the event orgaziner",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Event not found/No participant found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error found in DB",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/view/:id/benefits": {
            "get": {
                "description": "A user can view the benefits associated with an event that they own or was published",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "List the benefits of an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the event",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Event was not published yet and the user is not the orgaziner",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ParticipantInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
...
    },
...
}